refactor: reorganize architecture with DDD and SRP patterns
